requirements.txt.txt
psycopg2-binary
sqlalchemy
python-dotenv
pydantic
